log tomorrow rain

+++ weather4cast/weather.py
@@ -296,7 +296,7 @@ def get_rain_next_day(forecast_day, rain_limit):
         return False
     else:
         tomorrow_rain = weatherAPI.weekWeather[forecast_day + 1].rain    
-    rain_max = max(list(map(int, tomorrow_rain)))
+    rain_max = max(list(map(float, tomorrow_rain)))
     if rain_max > rain_limit:
         return True
     else:
